[grid] the way margins are chosen for baseline alignment
ascentForChild was determining the margin to use for the grid item's
ascent value using a combination of isDescentBaselineForChild
and isHorizontalAlignmentAxis. not really sure what this was supposed
to be doing but i think for determining the correct margin to use we
only need to consider the axis we are aligning in and the writing mode
of the grid.

for the column direction, that would be the "before" margin using the
grid's writing mode. for the row direction, that would be the "start"
margin using the grid's writing mode.