Replaced elements inside grid item with aspect ratio and padding grows indefinitely when resizing window. https://bugs.webkit.org/show_bug.cgi?id=283194 rdar://136048569
Reviewed by NOBODY (OOPS!).

Certain types of grid content with a replaced descendant that has
percent height and aspect ratio can trigger an unstable layout
in which the size of the content grows unexpectedly whenever the
inline constraints change. The following is a small snippet in
which this type of pathological behavior can occur:

<grid>
  <div class="aspect-ratio-and-padding">
    <div class="percent-height">
      <svg class="percent-height"></svg>
    </div>
  </div>
</grid>

Replacing the SVG with any sort of replaced element that does not have
an intrinsic size but does have an aspect ratio causes the same bug to
occur.

During track sizing, grid will query the grid item for its preferred
logical width. The preferred logical width of a replaced element without an
intrinsic logical width but with an intrinsic ratio and a non-auto value for its
logical height is computed from the ratio and logical height.

The replaced element has a percent-based height, so we will attempt to resolve its
value based off the height of the containing block. The containing block
of the replaced element is also percent-based, so we will do the same thing.
Its containing block is a grid item with an aspect ratio.

This lands us in availableLogicalHeightForPercentageComputation for the grid item
in which "shouldComputeLogicalHeightFromAspectRatio," returns true since the
width and height are both auto. We then try to compute the logical
height using the logical width off the renderer, which is a stale and
incorrect value for this context. Instead, let's just try to compute
the width of the grid item based on the constraints that are supposed to
be set by the grid via computeLogicalWidthInFragmentUsing.

* LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/aspect-ratio/replaced-elements-inside-grid-item-percent-height-and-padding-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/aspect-ratio/replaced-elements-inside-grid-item-percent-height-and-padding.html: Added.
* Source/WebCore/rendering/RenderBlock.cpp:
(WebCore::RenderBlock::availableLogicalHeightForPercentageComputation const):