[block-step-sizing] Start considering box's alignment when distributing extra space. https://bugs.webkit.org/show_bug.cgi?id=283585 rdar://140432622
Reviewed by NOBODY (OOPS!).

We currently do not implement the behavior for the block-step-align
property when it comes to determining how to split up the extra space
that is computed. However, the property does define an auto behavior
that refers to the box's self-alignment when adjusting the margin box,
which is currently the only type of adjustment we implement, so this
patch aims to start implementing that logic.

Essentially:
- If the box's self-alignment is start, we add the extra space to the
  block start side.
- If the box's self-alignment is end, we add the extra space to the block
  end side.
- Otherwise, split the extra space among the two edges like we already do.

The tests added are similar to the ones in 286969@main in that we start
providing support for some simple pieces of content. These are boxes
that have a specified inline size and do not have any border or padding.
Some of the tests specify a block size for the content, while some others
test when the box's block size is based on its content.

* LayoutTests/imported/w3c/web-platform-tests/css/css-rhythm/margin-distribution-align-items-simple-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-rhythm/margin-distribution-align-items-simple.html: Added.
Some simple content and how we distribute the space is based upon the
parent's align-items property.

* LayoutTests/imported/w3c/web-platform-tests/css/css-rhythm/margin-distribution-align-self-simple-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-rhythm/margin-distribution-align-self-simple.html: Added.
Some simple content and how we distribute the space is based upon the
box's align-self property.

* Source/WebCore/rendering/RenderBlockFlow.cpp:
(WebCore::RenderBlockFlow::layoutBlock):
(WebCore::RenderBlockFlow::distributeExtraBlockStepSizingSpaceToChild const):
(WebCore::RenderBlockFlow::layoutBlockChild):
* Source/WebCore/rendering/RenderBlockFlow.h: