[margin-trim][Invalidation] Block box descendants need invalidation when block container margin-trim is changed. https://bugs.webkit.org/show_bug.cgi?id=284954 rdar://124971386
Reviewed by NOBODY (OOPS!).

When margin-trim changes on a block container, this can have an impact on
block-level descendants since the effects of the property can propagate
to them. Currently, a margin-trim change requires layout on the block
container, but it does nothing to its descendants. This may mean we may
perform layout properly after this mutation occurs. Instead, we should
also propagate the damage to the subtree in the following ways:

1.) In LocalFrameViewLayoutContext::pushLayoutState, if we end up
pushing a new RenderLayoutState onto the stack, we should propagate the
potentially already existing margin trimming state. This requires
changing the constructor for RenderLayoutState to take in a bool for the
margin trimming state instead of just unconditionally initializing it to
false.

2.) In RenderBlockFlow::styleDidChange, traverse through the subtree and
setNeedsLayout on the descendants. If a descendant is a block box, then
continue to traverse through since it's possible that it and its
children, which are block-level boxes, are affected by the margin-trim
change from the block container. If it is not, then we should skip its
children.

* LayoutTests/imported/w3c/web-platform-tests/css/css-box/margin-trim/invalidation/block-container-block-start-child-with-border-no-trim-to-block-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-box/margin-trim/invalidation/block-container-block-start-child-with-border-no-trim-to-block.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-box/margin-trim/invalidation/block-container-block-start-self-collapsing-nested-no-trim-to-block-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-box/margin-trim/invalidation/block-container-block-start-self-collapsing-nested-no-trim-to-block.html: Added.
* Source/WebCore/page/LocalFrameViewLayoutContext.cpp:
(WebCore::LocalFrameViewLayoutContext::pushLayoutState):
* Source/WebCore/rendering/RenderBlockFlow.cpp:
(WebCore::RenderBlockFlow::styleDidChange):
* Source/WebCore/rendering/RenderIterator.h:
(WebCore::RenderTraversal::nextSkippingChildren):
(WebCore::RenderIterator<T>::traverseNextSkippingChildren):
* Source/WebCore/rendering/RenderLayoutState.cpp:
(WebCore::RenderLayoutState::RenderLayoutState):
* Source/WebCore/rendering/RenderLayoutState.h: