[Flex] Computing inner cross size of a flexbox with a definite cross size should consider box-sizing. https://bugs.webkit.org/show_bug.cgi?id=244697 rdar://99475341
Reviewed by NOBODY (OOPS!).

The flexbox spec states "If a single-line flex container has a
definite cross size, the automatic preferred outer cross size of
any stretched flex items is the flex container’s inner cross
size (clamped to the flex item’s min and max cross size) and is considered
definite."

We compute this cross size in computeCrossSizeForFlexItemUsingContainerCrossSize,
but we fail to take into consideration the box-sizing property. If the
flexbox has box-sizing: border-box specified on it then this will change
how we compute the inner cross size as we will need to take away any
border and padding from this fixed value.

* LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/flexbox-single-line-definite-cross-size-with-box-sizing-border-box-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/flexbox-single-line-definite-cross-size-with-box-sizing-border-box.html: Added.
* Source/WebCore/rendering/RenderFlexibleBox.cpp:
(WebCore::RenderFlexibleBox::computeCrossSizeForFlexItemUsingContainerCrossSize const):